requirements.txt
slack-bolt==1.18.0
python-dotenv==1.0.0
openai==1.3.0
python-dateutil==2.8.2
requests==2.31.0